move the shared authentictor login into a separate dependency.
