--- src/main/java/io/curity/identityserver/plugin/linkedin/authentication/PostDataParamBuilder.java
@@ -1,36 +1,0 @@
-/*
- *  Copyright 2017 Curity AB
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- */
-package io.curity.identityserver.plugin.linkedin.authentication;
-
-import org.apache.http.NameValuePair;
-import org.apache.http.message.BasicNameValuePair;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class PostDataParamBuilder {
-    private List<NameValuePair> pairs = new ArrayList<>();
-
-    public List<NameValuePair> getPairs() {
-        return pairs;
-    }
-
-    public PostDataParamBuilder addPair(String key, String value) {
-        pairs.add(new BasicNameValuePair(key, value));
-
-        return this;
-    }
-}
