--- src/main/java/io/curity/identityserver/plugin/linkedin/config/Constants.java
@@ -1,37 +1,0 @@
-/*
- *  Copyright 2017 Curity AB
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- */
-
-package io.curity.identityserver.plugin.linkedin.config;
-
-public class Constants {
-
-    public class Params {
-        public static final String PARAM_STATE = "state";
-        public static final String PARAM_CLIENT_ID = "client_id";
-        public static final String PARAM_SCOPE = "scope";
-        public static final String PARAM_RESPONSE_TYPE = "response_type";
-        public static final String PARAM_REDIRECT_URI = "redirect_uri";
-        public static final String PARAM_CLIENT_SECRET = "client_secret";
-        public static final String PARAM_CODE = "code";
-        public static final String PARAM_GRANT_TYPE = "grant_type";
-        public static final String PARAM_ACCESS_TOKEN = "access_token";
-        public static final String PARAM_URL = "url";
-        public static final String PARAM_GRANT_TYPE_AUTHORIZATION_CODE = "authorization_code";
-        public static final String PARAM_ID = "id";
-        public static final String BEARER = "Bearer";
-        public static final String PROFILE_REQUEST_URL = "siteStandardProfileRequest";
-    }
-}
